requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.35.0
peft>=0.6.0
accelerate>=0.24.0
datasets>=2.14.0
Pillow>=9.5.0
pandas>=1.5.0
numpy>=1.24.0
tqdm>=4.65.0
wandb>=0.15.0
safetensors>=0.4.0
xformers>=0.0.22
python-dotenv>=1.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0